Move GBD builder out of get_pop_data and pass HIV profile as vector
